Path: Refactor `Extensions` Gui code into independent module
Move the `Extensions` Gui code to independent module so access to other operations will be easier.
remove allaccessextensions
